+++ app/src/main/java/com/example/myapplication/data/HomeworkLogDao.kt
@@ -16,7 +16,7 @@ interface HomeworkLogDao {
     suspend fun updateHomeworkLog(homeworkLog: HomeworkLog)
 
     @Query("SELECT * FROM homework_logs WHERE studentId = :studentId ORDER BY loggedAt DESC")
-    fun getHomeworkLogsForStudent(studentId: Long): LiveData<List<HomeworkLog>>
+    fun getHomeworkLogsForStudent(studentId: Int): LiveData<List<HomeworkLog>>
 
     @Query("SELECT * FROM homework_logs ORDER BY loggedAt DESC")
     fun getAllHomeworkLogs(): LiveData<List<HomeworkLog>>
@@ -25,5 +25,5 @@ interface HomeworkLogDao {
     suspend fun deleteHomeworkLog(logId: Long)
 
     @Query("SELECT * FROM homework_logs WHERE studentId = :studentId ORDER BY loggedAt DESC LIMIT :limit")
-    fun getRecentHomeworkLogsForStudent(studentId: Long, limit: Int): LiveData<List<HomeworkLog>>
+    fun getRecentHomeworkLogsForStudent(studentId: Int, limit: Int): LiveData<List<HomeworkLog>>
 }
